refactor(service): replace redirect type with int64

+++ cli/downloader_creator.go
@@ -4,4 +4,4 @@ import (
 	afd "github.com/morozovcookie/afifiledownloader"
 )
 
-type DownloaderCreator func(isFollowRedirects bool, maxRedirects int) afd.DownloadFunc
+type DownloaderCreator func(isFollowRedirects bool, maxRedirects int64) afd.DownloadFunc
